Describing the new Feature
New plan time team! Let's add a method to the service where we can pass it the number of dinosaurs we want, how much security we want, and... it will take care of the rest! Let's examplify that!

How about function it_builds_enclosure_with_dinosaurs(). Inside, let's see, I'd like to be able to create a new Enclosure by saying $ enclosure = $ this->buildEnclosure(). We'll pass that the number of security systems we want - 1 - and the number of dinosaurs we want: 2. Then we can do some basic checks, like $ enclosure->shouldBeAnInstanceOf() to make sure an Enclosure is returned.

Oh, and very important! I want to make sure the new Enclosure has active security: $ enclosure->isSecurityActive(). Wait... but that's not auto-completing - I thought we added that method! Oh, it does exist, but it's private. We'll need to fix that in a minute. Anyways, use $ enclosure->isSecurityActive()->shouldReturn(true).

Simple enough! We're not asserting anything about the dinosaurs yet, but it's a good start. In Enclosure, make isSecurityActive() public: we've discovered that we do need to use this from outside of this class. And because it's public... and because I like to keep things organized, let's move it up above the private methods.