Implementing the Feature
With the tests red, let's write some code! Open EnclosureBuilderService and... fill in the method! I'll break the arguments onto multiple lines - they're gonna be a bit long - and advertise that this returns an Enclosure. Change the args to int  that defaults to 1... though it's entirely up to you if you want a default - and int $ numberOfDinosaurs = 3. Then, $ enclosure = new Enclosure() and we'll offload the real work to a private method: $ this->addSecuritySystems($ numberOfSecuritySystems, $ enclosure).

At the bottom, return $ enclosure - we'll worry about the dinosaurs in a minute. For the addSecuritySystems() method, I'm going to cheat and paste that in: you can find this function on the code block on this page. Make sure to re-type the y on Security and hit tab to auto-complete that and get the use statement on top.

It's nothing special: it takes in the $ numberOfSecuritySystems, does a for loop, chooses a random name and sets the 'is active' flag to true.

We're not adding any dinosaurs yet... which is fine... because we're not asserting anything about them yet either! We'll worry about that soon. Right now, run phpspec!
./vendor/bin/phpspec run
Green! This wasn't anything new, but now we have an awesome problem. In EnclosureBuilderService, we need to create some dinosaurs... but we are not going to create them by hand. Nope, we already have this beautiful DinosaurFactory that's great at growing and hatching Dinosaurs! That means that EnclosureBuilderService will need DinosaurFactory as a dependency. And that means, in order to finish the example, we are going to need to mock DinosaurFactory.