Dummy Objects Return Nothing!
Ok: now we need to enhance our example to describe the expected behavior for creating Dinosaurs. Basically, because we're passing '2' as the second argument, we would expect our DinosaurFactory to be called 2 times and for the final Enclosure to have 2 Dinosaurs. But... we haven't coded that yet. var_dump($ enclosure->getDinosaurs()). This will be an empty array, right? Try it:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:18

Ah, we were mostly right: it's an instance of the all-important Subject object. But if you look inside, the subject property is an empty array. Cool! Things are working like we expect... so far.

But here's where things get weird... or cool... or something: even if we added the code to EnclosureBuilderService to use DinosaurFactory to create the 2 Dinosaurs and add them to the Enclosure, the test would still fail! What?

Why? Because, when you create a 'dummy' object like DinosaurFactory it's not the real DinosaurFactory. And, by default, all of its methods return null and do nothing. It's a real... dummy. Right? So if we did write code here to use DinosaurFactory to create the dinosaurs... it wouldn't! It would return null and either the test would fail or, more likely, some code would blow up because it's expecting a Dinosaur object, not null.

Yep, if you simply tell prophecy to create a test double, it's referred to as a 'dummy' and... it does nothing. But, there are two things that we can do to make it more awesome. Let's talk about the first one next: controlling the return value when a method is called.