There are basically two big things you can do with a test double. First, you can add behavior: you can tell it exactly what value to return when a certain method is called, instead of returning null. Second, you add expectations. For example, you can assert that a certain method on DinosaurFactory should be called a certain number of times and even with some specific arguments.
Controlling Method Return Value

Right now, we need to control the return value of the growVelociraptor() method. Instead of returning null, which will probably explode when EnclosureBuilderService tries to add null to an Enclosure, we need it to return a Dinosaur object.

Check this out: create a  variable set to new Dinosaur() with Stegosaurus and false. And let's set its length to, how about, 6. Here is the key part: we want our DinosaurFactory dummy object to return this Dinosaur when somebody calls growVelociraptor(). I know... that's kind of confusing because this is not a velociraptor... but that proves my point! We can completely control how this behaves.

Do it with $ dinosaurFactory->growVelociraptor(). So, we pretend like $ dinosaurFactory is a real object and, just like normal with phpspec, we call methods on that object and pass in real arguments. Let's say that, whenever we use the EnclosureBuildersService, it will always grow a velociraptor of length 5. Then, to control the return value, say ->willReturn($ dino1).

That's it! Actually, we just did both things that I said you could do with a test double. First, by saying $ dinosaurFactory->growVelociraptor(5), we've added an assertion that if this method is called, it must be passed the argument 5. If any other value is passed, the test will fail. More on that later. And second, we've controlled the return value with ->willReturn().

There are a few other 'will' methods you can use to control the return value, and the most useful by far is just to say ->will() and pass that a callback function. That's super useful if a method is called multiple times and you need to return different values each time. More about that later too.

So... let's run the test!

 ./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:16
The test still passes... but we haven't actually added any new assertions yet.