Using the Object in EnclosureBuilderService
Let's get to work! Open EnclosureBuilderService. Here I'll hit Alt + Enter on $ dinosaurFactory and select 'Initialize Fields' to create and set that property. Down below, let's call a new method called addDinosaurs() and pass it the $ numberOfDinosaurs argument. To add that new method, I'll put my cursor on addDinosaur(), hit Alt + Enter and 'Add Method'.

Next, copy the inside of addSecuritySystems(), paste it here, then clear out the inside of the loop. Change the variable to $ numberOfDinosaurs and, very nicely, we can say $ enclosure->addDinosaur() and pass that $ this->dinosaurFactory->growVelociraptor(). And, remember: in the example we expected this to be called always with a length of 5.

Perfect! Move over and run the test again:
 ./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:16
It still passes. And, check out the wrapped object we're dumping: there are now two dinosaurs inside! And... interesting! They're actually the exact same Dinosaur. That makes sense: each time the growVelociraptor() method is called, our test double returns that same, one Dinosaur object.

This is cool because we can add a great assertion down here: $ enclosure->getDinosaurs()[0] - to get the first Dinosaur - ->shouldBe($ dino1). And, it's a little odd, but we can even check that the second Dinosaur is also this exact $ dino1 Dinosaur.

Try the test again:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:16
Still green! This is the first, great superpower of these test doubles, or dummy objects. By adding behavior to them, we can help control exactly what happens inside the class we're testing. And, it often allows us to have very specific assertions.