Returning a Different Object Each Time
To make this a bit more realistic, copy $ dino1 and make a new $ dino2 - how about a Baby Stegosaurus with length 2. Adorable. I want to change the $ dinosaurFactory test double so that it returns $ dino1 the first time growVelociraptor() is called and $ dino2 the second time it's called... which is a bit more how things would work in the real world. How can we do this? By passing a second argument to willReturn().

That's it! Down below, the second object will now be . Try the tests one last time.
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:16

Green! Geez - we didn't manage to break anything in this video - we gotta try harder! By the way, when you control the return value of a test double, it's then called a 'Stub'... which is probably not that important to know, except for impressing other programmers at a party.

Next: let's talk a little bit more about this '5' argument. As it turns out, there are a lot of interesting things you can do with this argument. Like, what if EnclosureBuilderService always calls this method and passes a random number? What should we put here? Or, what if it's called multiple times with different arguments each time? Let's jump on that!