In our spec, we said that, when we call buildEnclosure(), we expect it to call growVelociraptor() on DinosaurFactory and pass it the exact argument 5. There was no super cool or secret reason for that: it's just how we decided to make this method work: all dinosaurs would have this same length. And when it calls growVelociraptor() with an argument of 5, we said that it should return  the first time it's called and  the second time.
Calling a Stubbed Method with a Different Argument

But... what if it weren't that simple? Go into the service. Making every dinosaur the same size isn't very realistic. No, let's make it more interesting - let's make the length 5 + $ i. So 5, then 6, 7 and so on.

What will phpspec think of this? Our example still says that we expect this method to always be called with the argument 5. Well... let's find out!
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:19
It fails! And check this out! It says:

  Unexpected method call on DinosaurFactory, growVelociraptor(6). Expected was growVelociraptor(5)

The way this fails is the really important part. When you call willReturn(), this is called a method promise. You're making a promise that the growVelociraptor() method will return this $ dino1 value and then this $ dino2 value. As soon as you apply even one promise to one method on a dummy object, you must apply a promise to every single call.

This can be a little tricky to understand at first. When we say $ dinosaurFactory->growVelociprator(5)->willReturn(), we're really saying:

  Hey phpspec! When growVelociraptor() is called and passed 5 as an argument, you should return $ dino1  and $ dino2. If any other value is passed, this promise doesn't apply.

In other words, the simplest way to fix this would be to add a promise for every argument that we'll pass - like $ dinosaurFactory->growVelociraptor(6)->willReturn(), then 7, 8, 9 - however many you need for that example. Whenever growVelociraptor() is called, prophecy goes down all of the promises for that method and finds the one that fits best. If none are found... but at least one is specified... error!