Argument::any()
Of course... creating a promise for every possible argument is... a bit nuts. So what's the real fix? Remove the specific argument 5 and replace it with a special Argument class from prophecy and pass it any(). Try phpspec again:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17
We're back! This says:
  Yo phpspec! It's me again, Ryan. So, when growVelociraptor() is called with any argument, here's what you should return.

Thanks to this, every time we call growVelociraptor() - regardless of the arguments passed to it - this promise will be matched.