But sometimes, I want to be a bit more specific. For example, we don't really want growVelociraptor() to be called with any argument: it should be an integer at least! If you really want to make sure that's happening correctly, you can use Argument::type() and pass integer.
This should work for us... so let's try it!
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17

All green! But now change this to string. Try it again:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:19
Yes! Unexpected method call on growVelociraptor(5). When the integer 5 is passed, phpspec says:

  Hey, somebody is calling growVelociraptor() with an argument that doesn't match any of the promises for this method. That's probably not expected, so I'm gonna go ahead and explode. Cheers!

The most flexible Argument is called that(). I won't show an example, but Argument::that() allows you to pass a callback function. It passes your callback the argument and then you can do whatever logic you need to determine if this is something you're expecting or not.

Let's change this back to type('integer'). And here is where things get truly interesting. We know that growVelociraptor() is going to be called two times. And so, to take super crazy control of things, we can create two separate method promises to handle each one. If you don't fully understand how this Argument stuff works yet, you will next.