The growVelociraptor() method is called two times. And thanks to the two arguments to willReturn(), $ dino1 will be returned first, then $ dino2. But we can control this much more than we are now.
Check it out: copy the beginning of the promise and paste it below. Replace the argument with 5. Then say ->willReturn($ dino2). For the first promise, remove $ dino2: this will now return $ dino1 every time its matched.

So... what do you think will happen now? Down here, I'm saying that when growVelociraptor() is called with exactly the number 5, return $ dino2. But up here I'm saying: if growVelociraptor() is called with any integer, return $ dino1. So... when growVelociraptor() is called with the number 5, what's going to be returned? $ dino1? $ dino2? A mystery !? Does the order of these promises matter? Or something else? Is there a 9th planet in the solar system that we haven't discovered yet?

I don't know! Let's find out! Or at least, answer a few of these questions. Try phpspec:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:19
It fails: expected Dinosaur but got Dinosaur. Wow... umm... that's not very helpful. Re-run with -v:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:19 -v
Ah! Now we can see that, on line 38, so when we're checking the first dinosaur, we expect it to be our big Stegosaurus, but it was actually our Baby Stegosaurus.

Arguments Matching is by 'Most Specific'

This is really cool! Each time we call growVelociraptor(), it looks at all of these method promises and finds the most specific one. When the argument is 5, it matches both promises: it is 5, but it's also an integer. But because the second is more specific, that promise is used and it returns $ dino2.

So, the first time we go through the loop, 5 is the argument and we return $ dino2. The second time, the argument is 6 and only matches the first promise. So, $ dino1 is returned.