So, the first time we go through the loop, 5 is the argument and we return $ dino2. The second time, the argument is 6 and only matches the first promise. So, $ dino1 is returned.
This means that $ dino2 is now the first item and $ dino1 is the second. How cool is that?! That's full, beautiful control.

Let's take off the -v option and run the tests:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17
Got it! To prove how this works, we can even re-order the two promises. Yep, makes no difference.
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17
So, whenever a method is called on a stub, prophecy will look at all the promises for that method and find the best one. And if none are found... but at least one exists, error!