Let's change this back to just one promise: we don't really need to be this specific. Delete the more specific call and make the other one return $ dino1 and then $ dino2 just like before. Also update the asserts to go back to the original way. Double-check that phpspec is happy: ./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17
Woohoo! Arguments are a great way to very specifically control what values are returned based on the input. And to make sure your methods are being called with the arguments you expect.

Next: instead of controlling the return value of methods, let's talk about adding expectations to our test doubles - like making sure that a method was called an exact number of times. Yep, it's mocks and spies time