Hey! We are experts when it comes to controlling the return value of any method on a test double. Here, we told it that when growVelociraptor() is called with any integer argument, return $ dino1 the first time and $ dino2 the second time.
By doing this, we also added some expectations. Because, as soon as you control the return value of one method call, then you must control the behavior of all method calls. And so if a different method is called or different arguments are passed to this method, the tests will fail.

But what this does not guarantee is that growVelociraptor() was actually called... or how many times it was called. Nope, we're saying, if it's called, it must be called with this argument... and here's what to return. But technically, if it were called zero times, this part would not fail!

And that's what I want to talk about now: sometimes it is super important to make sure a method was called... or was called exactly three times... or was not called! For example, what if we wanted to be absolutely sure that growVelociraptor() was called exactly two times? We can do that, and it starts the same way: $ dinosaurFactory->growVelociraptor(Argument::type('integer')). But then, instead of willReturn(), use shouldBeCalledTimes(2).

This feels familiar... because it's exactly like what we've been doing! It looks like a matcher! It's identical to $ this->shoudHaveType(), for example.
Now that we're asserting that this method should be called twice, run phpspec:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17
Nice! It passes! What does it look like to fail? Change this to 3... and try it again:
 ./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:17
  Perfect: Expected exactly 3 calls that match growVelociraptor() with type integer, but 2 were made.