Now that this example is executing really well... ./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:19 let's go back and run all of the examples in this spec class.
 ./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php
Ah, oh yea! it_is_initializable() is still failing: too few arguments to the constructor: zero passed, expected one. That makes sense: we added $ this->beConstructedWith() to our it_builds_enclosures_with_dinosaurs() example, but not to this example.

The easiest solution is just to... ya know... duplicate it! But... it does kind of make me think: it would be nice if we could call a method before each example was executed - a method that could, sort of, set some things up. After all, we're going to need to call $ this->beConstructedWith() in every single example.