Call a method before each example was executed - a method that could, sort of, set some things up, by creating a function called let(). Yep! let() will be called once before each example function is executed. So, let() then it_is_initializable(), then let() again and it_builds_enclosures_with_dinosaurs().
Inside let(), we can do the exact same things as our normal methods... meaning, if we need a test double, we can add a DinosaurFactory $ dinosaurFactory argument. And this allows us to move the beConstructedWith() call up to let().

Ok, let's look at this: let() will be called first, and will be passed the $ dinosaurFactory test double. But then, down in the example, we need to make sure that we get that exact same DinosaurFactory test double object because that is the object we need to add behavior to.

And that is exactly how this will work. But, it's not just because these are both DinosaurFactory objects. Nope. phpspec matches by the argument name: because the argument is called $ dinosaurFactory in let() and because it has the same name below, phpspec knows to pass the same object... instead of creating a brand-new test double.

So... unless we've mucked something up, this should make everything happy! Try it:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php:19
Woohoo