Using Mocks/Spies to Guarantee Saving to the Database
Let's do one last thing before you all go off and take over the world with your new phpspec knowledge. One of the things I want my EnclosureBuilderService to do is to save the new EnclosureBuilder to the database after it finishes building it. Now, obviously, we don't have a database in this application... but who cares? We can fake it!

Open up the tutorial/ directory: you'll have it if you downloaded the course code. In the Service/ directory there is an EntityManagerInterface.php file. Copy that and put it into your Service/ directory.

If you're a Doctrine user, this will look familiar. But, no, I'm not recommending that you actually create or move the EntityManagerInterface into your app like this. We're adding this interface as a convenient way to 'pretend' like Doctrine exists in our app. This interface looks just like the one from Doctrine, with the same persist() and flush() methods.