Guaranteeing the Object is Saved
Back in the example, I want to describe that persist() and flush() should be called on the EntityManagerInterface when we call buildEnclosure(): I want to guarantee that I didn't forget to save this to the database. And that means that our EnclosureBuilderService will have a second dependency. In let(), add a second argument: EntityManagerInterface $ entityManager. Pass that as the second argument to beConstructedWith()

Then, down in the actual example, we want to assert that persist() and flush() were called on it. Get that same test double here by also saying EntityManagerInterface $ entityManager.

At this point, we can use the mock or spy functionality... it makes no difference. I'll do it as a spy. Start with $ entityManager->persist(). This should be passed an Enclosure object. So let's say Argument::type() with Enclosure::class. Then, ->shouldHaveBeenCalled().

Repeat this with flush(), except that flush() doesn't take any arguments.
This time, we're not giving the test double any behavior: it's a pure spy. Try it:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php
Woo! It fails:

no calls have been made that match persist() with type Enclosure, but expected at least one.