Open EnclosureBuilderService and let's code that up. Start by adding the EntityManagerInterface  argument. I'll hit Alt + Enter to create that property and set it
Finally, at the bottom of the method, $ this->entityManager->persist($ enclosure) and $ this->entityManager->flush().

This looks great. But let's run phpspec one last time to be sure:
./vendor/bin/phpspec run spec/Service/EnclosureBuilderServiceSpec.php
It passes! Try the entire test suite:
./vendor/bin/phpspec run
Hey! We didn't break anything!

Friends! That's it. phpspec is wonderful tool to help you write unit tests... but really focus on the design of your class. Yes, you do need to get used to a lot of the magic it does. But once you embrace the magic, the experience is wonderful... and the code generation isn't too bad either.

One word of warning that I always like to give people is this: just because you have a wonderful testing tool like phpspec, it doesn't mean you need to test every single thing. For example, in DinosaurSpec, we're doing a lot of testing on the getter and setter methods. You can do that... but I think it's overkill. Think of testing less as an 'all-or-nothing' sort of thing, and more of a priority system: make it a high priority to test, or describe the classes that have a lot of complexity or that scare you.

Ok, get out there, describe some great classes, and we'll seeya next time!